+++ spec/detectors/delegation_n_plus_one_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe EagerEye::Detectors::DelegationNPlusOne do
 
   describe "#detect" do
     def parse(source)
-      Parser::CurrentRuby.parse(source)
+      EagerEye::SourceParser.parse(source)
     end
 
     context "with local delegate declarations" do