+++ spec/detectors/loop_association_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe EagerEye::Detectors::LoopAssociation do
 
   describe "#detect" do
     def parse(source)
-      Parser::CurrentRuby.parse(source)
+      EagerEye::SourceParser.parse(source)
     end
 
     context "with each iteration" do